Correcion en index.py
