+++ app/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from flask_bootstrap import Bootstrap
 from .config import Config
-from .routes.perfil import perfil_ruta
+from .routes.index import perfil_ruta
 
 def create_app():
     app = Flask(__name__)
