+++ app/templates/index2.html
@@ -3,8 +3,11 @@
 {% block content %}
 <h1>Lista de personajes</h1>
     {% for personaje in lista %}
-    <img src="{{personaje.image}}">
+    <a href="{{url_for('perfil_ruta.perfil')}}">
+        <img src="{{personaje.image}}">
+    </a>
     <ul>
+        
         name: {{personaje.name}}<br>
         status: {{personaje.status}}<br>
         species: {{personaje.species}}<br>
@@ -12,6 +15,8 @@
         gender: {{personaje.gender}}<br>
         origin: {{personaje.origin}}<br>
         location: {{personaje.location}}<br>
+       
+                      
         <br>
         <br>
