+++ app/templates/perfil.html
@@ -1,0 +1,27 @@
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+
+    <title>profile</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/water.css@2/out/water.css">
+</head>
+
+<body>
+    
+    <img src="https://rickandmortyapi.com/api/character/avatar/1.jpeg">
+    <br>
+    <ul>
+        
+        name: <br>
+        status: <br>
+        species: <br>
+        type: <br>
+        gender: <br>
+        origin: <br>
+        location: <br>
+        <br>
+        <br>
+</body>
+
+</html>
